+++ include/power_source_store.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "power_source.h"
+
+app_config::PowerSource loadPowerSourcePreference();
+void savePowerSourcePreference(app_config::PowerSource source);